Fix prepare_ functions with related
If the prepare_ function on a document use a models
from the Meta.related_models it doesn't manage when
the related_models is deleted. Add a new prepare_X_with_related
that take another param `related_instance_to_ignore` which represent
the related_model being deleted. If this models instance is used in
the prepare_ function we now avoid to return it.

See #64